# ui/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "",
    "version": ""
  },
  "paths": {
...
    "/config": {
      "get": {
        "operationId": "get_config",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppConfig"
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "patch_config",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppConfig"
                }
              }
            }
          }
        }
      }
    },
    "/config/providers/{id}/secret": {
      "put": {
        "summary": "Save (or overwrite) the keyring secret for a provider. Creates the\nprovider entry in `config.providers` if it didn't exist. `PUT` because\nsetting the secret is idempotent for the same body.",
        "operationId": "set_provider_secret",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Provider id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderSecretRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": ""
          }
        }
      },
      "delete": {
        "summary": "Clear a provider's keyring secret. The provider entry itself is kept.",
        "operationId": "clear_provider_secret",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Provider id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AppConfig": {
        "type": "object",
        "properties": {
          "data": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/DataConfig"
              }
            ],
            "default": {
              "path": "/Users/w/Library/Application Support/Koharu"
            }
          },
          "http": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/HttpConfig"
              }
            ],
            "default": {
              "connect_timeout": 20,
              "max_retries": 3,
              "read_timeout": 300
            }
          },
          "huggingface": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/HuggingFaceConfig"
              }
            ],
            "default": {}
          },
          "pipeline": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/PipelineConfig"
              }
            ],
            "default": {
              "bubble_segmenter": "speech-bubble-segmentation",
              "detector": "koharu-yolo26s",
              "font_detector": "yuzumarker-font-detection",
              "inpainter": "lama-manga",
              "ocr": "paddle-ocr-vl-1.6",
              "renderer": "koharu-renderer",
              "segmenter": "comic-text-detector-seg",
              "translator": "llm"
            }
          },
          "providers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProviderConfig"
            },
            "default": []
          }
        }
      },
...
      "ConfigPatch": {
        "type": "object",
        "description": "Sparse patch for `koharu_app::AppConfig`. Missing fields mean \"leave\nas-is\". The `providers` field, if present, replaces the whole provider\nlist — we do not merge by id because ordering is meaningful.",
        "properties": {
          "data": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DataConfigPatch"
              }
            ]
          },
          "http": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/HttpConfigPatch"
              }
            ]
          },
          "huggingface": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/HuggingFaceConfigPatch"
              }
            ]
          },
          "pipeline": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PipelineConfigPatch"
              }
            ]
          },
          "providers": {
            "type": ["array", "null"],
            "items": {
              "$ref": "#/components/schemas/ProviderPatch"
            },
            "description": "If present, replaces the entire list. Api_key values of `\"[REDACTED]\"`\nare interpreted as \"leave the existing secret alone\"."
          }
        }
      },
...
      "DataConfig": {
        "type": "object",
        "required": ["path"],
        "properties": {
          "path": {
            "type": "string"
          }
        }
      },
      "DataConfigPatch": {
        "type": "object",
        "properties": {
          "path": {
            "type": ["string", "null"]
          }
        }
      },
...
      "HttpConfig": {
        "type": "object",
        "properties": {
          "connect_timeout": {
            "type": "integer",
            "format": "int64",
            "default": 20,
            "minimum": 0
          },
          "max_retries": {
            "type": "integer",
            "format": "int32",
            "default": 3,
            "minimum": 0
          },
          "read_timeout": {
            "type": "integer",
            "format": "int64",
            "default": 300,
            "minimum": 0
          }
        }
      },
      "HttpConfigPatch": {
        "type": "object",
        "properties": {
          "connectTimeout": {
            "type": ["integer", "null"],
            "format": "int64",
            "minimum": 0
          },
          "maxRetries": {
            "type": ["integer", "null"],
            "format": "int32",
            "minimum": 0
          },
          "readTimeout": {
            "type": ["integer", "null"],
            "format": "int64",
            "minimum": 0
          }
        }
      },
      "HuggingFaceConfig": {
        "type": "object",
        "description": "HuggingFace access, for the gated model repositories.",
        "properties": {
          "token": {
            "type": ["string", "null"],
            "description": "Populated from credential storage on `load()`, never written to\nconfig.toml. Serializes as `\"[REDACTED]\"` in API responses.\n\nOnly some model repositories need this. `deepghs/AnimeText_yolo` is\n`gated: auto`: the terms have to be accepted on the repo page and the\nrequest has to be authenticated. When unset, koharu still honors\n`HF_TOKEN` and the `hf auth login` file.",
            "default": null
          }
        }
      },
      "HuggingFaceConfigPatch": {
        "type": "object",
        "properties": {
          "token": {
            "type": ["string", "null"],
            "description": "`\"[REDACTED]\"` → keep existing keyring secret; empty → clear; otherwise save."
          }
        }
      },
...
      "PipelineConfig": {
        "type": "object",
        "description": "Engine selection for each pipeline stage.\nValues are engine IDs (e.g. \"koharu-yolo26s\", \"comic-text-detector-seg\").\nEmpty string means use default.",
        "properties": {
          "bubble_segmenter": {
            "type": "string",
            "default": "speech-bubble-segmentation"
          },
          "detector": {
            "type": "string",
            "default": "koharu-yolo26s"
          },
          "font_detector": {
            "type": "string",
            "default": "yuzumarker-font-detection"
          },
          "inpainter": {
            "type": "string",
            "default": "lama-manga"
          },
          "ocr": {
            "type": "string",
            "default": "paddle-ocr-vl-1.6"
          },
          "renderer": {
            "type": "string",
            "default": "koharu-renderer"
          },
          "segmenter": {
            "type": "string",
            "default": "comic-text-detector-seg"
          },
          "translator": {
            "type": "string",
            "default": "llm"
          }
        }
      },
      "PipelineConfigPatch": {
        "type": "object",
        "properties": {
          "bubbleSegmenter": {
            "type": ["string", "null"]
          },
          "detector": {
            "type": ["string", "null"]
          },
          "fontDetector": {
            "type": ["string", "null"]
          },
          "inpainter": {
            "type": ["string", "null"]
          },
          "ocr": {
            "type": ["string", "null"]
          },
          "renderer": {
            "type": ["string", "null"]
          },
          "segmenter": {
            "type": ["string", "null"]
          },
          "translator": {
            "type": ["string", "null"]
          }
        }
      },
...
      "ProviderConfig": {
        "type": "object",
        "required": ["id"],
        "properties": {
          "api_key": {
            "type": ["string", "null"],
            "description": "Populated from credential storage on `load()`, never written to config.toml.\nSerializes as `\"[REDACTED]\"` in API responses."
          },
          "base_url": {
            "type": ["string", "null"]
          },
          "id": {
            "type": "string"
          }
        }
      },
      "ProviderPatch": {
        "type": "object",
        "required": ["id"],
        "properties": {
          "apiKey": {
            "type": ["string", "null"],
            "description": "`\"[REDACTED]\"` → keep existing keyring secret; empty → clear; otherwise save."
          },
          "baseUrl": {
            "type": ["string", "null"]
          },
          "id": {
            "type": "string"
          }
        }
      },
      "ProviderSecretRequest": {
        "type": "object",
        "required": ["secret"],
        "properties": {
          "secret": {
            "type": "string"
          }
        }
      },
...
    }
  }
}